Add files via upload
Here is my first intermediate project from Karpov Courses on Data analysis. The main purpose - analize data from online store and give responses on key questions.
Also it is needed to add that there is no A\B, A\A tests and other statistical methods applied, just because it was learned after the project's deadline.